backend/pyproject.toml
[project]
name = "cramwell"
version = "0.3.1"
description = "An OSS and LlamaCloud-backed alternative to NotebookLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.104.1",
  "fastmcp>=2.9.2",
  "ffprobe>=0.5",
  "jinja2>=3.1.0",
  "llama-index-core>=0.12.44",
  "llama-index-llms-openai>=0.4.7",
  "llama-index-tools-mcp>=0.2.5",
  "markdown-analysis>=0.1.5",
  "mypy>=1.16.1",
  "opentelemetry-exporter-otlp-proto-http>=1.34.1",
  "pinecone>=3.0.0",
  "pre-commit>=4.2.0",
  "psycopg2-binary>=2.9.10",
  "pytest>=8.4.1",
  "pytest-asyncio>=1.0.0",
  "python-dotenv>=1.1.1",
  "uvicorn>=0.24.0",
  "supabase>=2.6.0",
  "beautifulsoup4>=4.12.0",
  "docling>=0.1.0",
  "PyPDF2>=3.0.0",
  "python-docx>=1.1.0",
  "psutil>=5.9.0",
  "aiohttp>=3.9.0"
]

[tool.mypy]
disable_error_code = ["import-not-found", "import-untyped"]

[tool.pytest.ini_options]
pythonpath = ["src"]
